// frontend/src/app/generated/model/taskDTO.ts
/**
 * xRM
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ResponsibleDTO } from './responsibleDTO';
import { CustomerDTO } from './customerDTO';


export interface TaskDTO { 
    id?: number;
    subject?: string;
    type?: string;
    description?: string;
    time?: number;
    responsible?: ResponsibleDTO;
    partner?: CustomerDTO;
}